+++ Thunderstore/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### v1.0.2
+
+- Add affect vanilla, affect modded configs
+- Add a blacklist config, where you can write names of hazards you wanna disable
+- Add a HazardNames list, this not a config, this is just for you to know what hazard names is in your mod pack. You reset to default to refresh list data.
+
 ### v1.0.1
 
 - Add customVisibleRange, where you can custom the visible range for every hazards by using their prefab name